+++ src/test/java/com/alibaba/json/bvt/serializer/enum_/EnumFieldsTest6.java
@@ -1,7 +1,4 @@
-package com.alibaba.json.bvt.serializer;
-
-import java.util.ArrayList;
-import java.util.List;
+package com.alibaba.json.bvt.serializer.enum_;
 
 import org.junit.Assert;
 
@@ -10,18 +7,17 @@ import com.alibaba.fastjson.serializer.SerializerFeature;
 
 import junit.framework.TestCase;
 
-public class EnumFieldsTest3 extends TestCase {
+public class EnumFieldsTest6 extends TestCase {
     public void test_enum() throws Exception {
         Model model = new Model();
-        model.types.add(Type.A);
-        model.types.add(null);
+        model.types = new Object[]{Type.A, null};
         
         String text = JSON.toJSONString(model, SerializerFeature.WriteMapNullValue);
         Assert.assertEquals("{\"types\":[\"A\",null]}", text);
     }
     
     public static class Model {
-        public List<Type> types = new ArrayList<Type>();
+        public Object[] types;
     }
 
     private static enum Type {
