+++ src/test/java/com/alibaba/json/bvt/serializer/enum_/EnumTest2.java
@@ -1,4 +1,4 @@
-package com.alibaba.json.bvt.serializer;
+package com.alibaba.json.bvt.serializer.enum_;
 
 import java.util.Date;
