+++ src/test/java/com/alibaba/json/bvt/serializer/enum_/EnumTest3.java
@@ -8,7 +8,7 @@
  * qzhanbo@yiji.com 2015-03-01 00:55 创建
  *
  */
-package com.alibaba.json.bvt.serializer;
+package com.alibaba.json.bvt.serializer.enum_;
 
 import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.serializer.SerializeWriter;
